+++ DotaHeroes/API/Events/EventArgs/Hero/HeroAttackedEventArgs.cs
@@ -13,7 +13,7 @@ namespace DotaHeroes.API.Events.EventArgs.Hero
 
         public Features.Hero Target { get; }
 
-        public decimal Damage { get; set; }
+        public double Damage { get; set; }
 
         public DamageType DamageType { get; set; }
 
@@ -24,7 +24,7 @@ namespace DotaHeroes.API.Events.EventArgs.Hero
         /// <param name="target"><inheritdoc cref="Target" /></param>
         /// <param name="damage"><inheritdoc cref="Damage" /></param>
         /// <param name="damageType"><inheritdoc cref="DamageType" /></param>
-        public HeroAttackedEventArgs(Features.Hero hero, Features.Hero target, decimal damage, DamageType damageType)
+        public HeroAttackedEventArgs(Features.Hero hero, Features.Hero target, double damage, DamageType damageType)
         {
             Hero = hero;
             Target = target;
