+++ DotaHeroes/API/Events/EventArgs/Hero/HeroTakedDamageEventArgs.cs
@@ -13,7 +13,7 @@ namespace DotaHeroes.API.Events.EventArgs.Hero
 
         public Features.Hero Attacker { get; }
 
-        public decimal Damage { get; set; }
+        public double Damage { get; set; }
 
         public DamageType DamageType { get; set; }
 
@@ -24,7 +24,7 @@ namespace DotaHeroes.API.Events.EventArgs.Hero
         /// <param name="attacker"><inheritdoc cref="Attacker" /></param>
         /// <param name="damage"><inheritdoc cref="Damage" /></param>
         /// <param name="damageType"><inheritdoc cref="DamageType" /></param>
-        public HeroTakedDamageEventArgs(Features.Hero hero, Features.Hero attacker, decimal damage, DamageType damageType)
+        public HeroTakedDamageEventArgs(Features.Hero hero, Features.Hero attacker, double damage, DamageType damageType)
         {
             Hero = hero;
             Attacker = attacker;
